# Conflicts:
#	src/main/java/com/ybe/ybe_toyproject3/domain/itinerary/service/ItineraryService.java
#	src/main/java/com/ybe/ybe_toyproject3/domain/trip/model/Trip.java
#	src/main/java/com/ybe/ybe_toyproject3/domain/trip/service/TripService.java
#	src/main/java/com/ybe/ybe_toyproject3/domain/user/repository/UserRepository.java

+++ src/main/java/com/ybe/ybe_toyproject3/domain/itinerary/service/ItineraryService.java
@@ -17,6 +17,7 @@ import com.ybe.ybe_toyproject3.domain.trip.model.Trip;
 import com.ybe.ybe_toyproject3.domain.trip.repository.TripRepository;
 import com.ybe.ybe_toyproject3.global.component.KakaoApiComponent;
 
+import com.ybe.ybe_toyproject3.global.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
@@ -39,6 +40,13 @@ public class ItineraryService {
 
     @Transactional
     public ItineraryCreateResponse createItinerary(Long tripId, ItineraryCreateRequest itineraryCreateRequest)  {
+        Trip trip = validateTrip(tripId);
+
+        Long currentUserId = SecurityUtil.getCurrentUserId();
+        Long tripUserId = trip.getUser().getId();
+
+        checkItineraryAuthorization(currentUserId, tripUserId);
+
         String locationName = createLocation(itineraryCreateRequest.getPlaceName());
         Location location = new Location(locationName);
 
@@ -74,6 +82,13 @@ public class ItineraryService {
             Long itineraryId, ItineraryUpdateRequest request,
             Long tripId
     ) {
+        Itinerary itinerary = validateItinerary(itineraryId);
+
+        Long currentUserId = SecurityUtil.getCurrentUserId();
+        Long tripUserId = itinerary.getTrip().getUser().getId();
+
+        checkItineraryAuthorization(currentUserId, tripUserId);
+
         String locationName = createLocation(request.getPlaceName());
 
         Itinerary itinerary = validateItinerary(itineraryId);
@@ -88,14 +103,23 @@ public class ItineraryService {
     }
 
     @Transactional
-    public String deleteItinerary(Long itineraryId, Long tripId) {
-        validateItinerary(itineraryId);
+    public String deleteItinerary(Long itineraryId) {
+        Itinerary itinerary = validateItinerary(itineraryId);
 
-        Itinerary itinerary = itineraryRepository.findById(itineraryId).get();
+        Long currentUserId = SecurityUtil.getCurrentUserId();
+        Long tripUserId = itinerary.getTrip().getUser().getId();
+
+        checkItineraryAuthorization(currentUserId, tripUserId);
 
         itineraryRepository.delete(itinerary);
 
         return itineraryId+"번 여정이 삭제되었습니다.";
+    }
+
+    private void checkItineraryAuthorization(Long currentUserId, Long tripUserId) {
+        if (!currentUserId.equals(tripUserId)) {
+            throw new RuntimeException("해당 여행에 대한 여정을 조작할 수 있는 권한이 없습니다.");
+        }
     }
 
     private Trip validateTrip(Long tripId) {
